requirements.txt
Flask==2.0.1
Pillow==8.3.1
pywebview==3.6.3
pyinstaller==4.5.1
